applied singleton pattern to lazily create RocksDBStateBackend

+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/execution/ExecutionContext.java
@@ -108,8 +108,6 @@ public class ExecutionContext implements DynamicMetricsProvider {
 
     private InternalSerializationService serializationService;
 
-    private RocksDBStateBackend rocksDBStateBackend;
-
 
     public ExecutionContext(NodeEngine nodeEngine, TaskletExecutionService taskletExecService,
                             long jobId, long executionId, Address coordinator, Set<Address> participants) {
@@ -138,9 +136,8 @@ public class ExecutionContext implements DynamicMetricsProvider {
         JetService jetService = nodeEngine.getService(JetService.SERVICE_NAME);
         serializationService = jetService.createSerializationService(jobConfig.getSerializerConfigs());
         metricsEnabled = jobConfig.isMetricsEnabled() && nodeEngine.getConfig().getMetricsConfig().isEnabled();
-        rocksDBStateBackend = new RocksDBStateBackend(serializationService);
         plan.initialize(nodeEngine, jobId, executionId, snapshotContext,
-                tempDirectories, serializationService, rocksDBStateBackend);
+                tempDirectories, serializationService);
         snapshotContext.initTaskletCount(plan.getProcessorTaskletCount(), plan.getStoreSnapshotTaskletCount(),
                 plan.getHigherPriorityVertexCount());
         receiverMap = unmodifiableMap(plan.getReceiverMap());
@@ -223,8 +220,11 @@ public class ExecutionContext implements DynamicMetricsProvider {
             serializationService.dispose();
         }
 
-        rocksDBStateBackend.deleteKeyValueStore();
-        IOUtil.delete(rocksDBStateBackend.getDirectory());
+        if (RocksDBStateBackend.exists()) {
+            RocksDBStateBackend rocksDBStateBackend = RocksDBStateBackend.getInstance(serializationService);
+            rocksDBStateBackend.deleteKeyValueStore();
+            IOUtil.delete(rocksDBStateBackend.getDirectory());
+        }
     }
 
     /**
